requirements.txt
asgiref==3.11.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2.10
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
fonttools==4.61.1
gunicorn==24.0.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
packaging>=24.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
